refactor(domain) : enum 클래스 경로 변경
constant -> domain
